✨ PoC as WebAuthnClientBackend
